requirements.txt
alembic==1.14.1
asttokens==3.0.0
backcall==0.2.0
beautifulsoup4==4.13.4
blinker==1.8.2
click==8.1.3
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
Flask==3.0.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.5
rich==14.0.0
shellingham==1.5.4
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typer==0.16.0
typing_extensions==4.13.2
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
